requirements.txt
gradio==4.44.0
requests
numpy
pypdf
sentence-transformers
faiss-cpu
langchain-text-splitters
datasets
